Skip tests for missing modules and multiprocessing fork
The installation of jax and pytorch will be skipped on unsupported
platforms now. The tests using jax or pytorch will be skipped, if
jax or pytorch is not installed, respectively.

The test for shared memory, where different start methods ('spawn'
and 'fork') are tested, will be skipped for start method 'fork' if
it is not available, which is the case on windows.